requirements.txt
pytest==5.3.0
requests==2.22.0
azure-functions
